Add route and controller midtrans * success response order regular reduce qty on hand and preorder increment order received. * expire send mail to billing_email.

+++ modules/ecommerce/src/Http/Controllers/Api/CheckoutApiController.php
@@ -108,6 +108,17 @@ class CheckoutApiController extends Controller
                         ];
                         $subtotal       = intval($value->qty) * intval($value->price);
                         $total_amount += intval($subtotal);
+                        
+                        $qty_reduced = intval($value->productVariant->quantity_on_hand) - intval($value->qty);
+                        
+                        if ($qty_reduced <= 0 ) {
+                            
+                            return response()->json([
+                                'data' => 'Out of stock ',
+                                'status' => 462
+                            ]);
+                        }
+
                         $value->delete();
                     }
                 }
